toybox_core/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "toybox_core"
version = "0.0.1"
description = "toy robotics stack core"
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    {name = "Wes Fletcher", email = "please@donot.com"},
]
requires-python = ">=3.8.2"
dependencies = [
    "grpcio", # gRPC
    "grpcio-tools",
    "toybox_msgs",
]

[project.urls]
"Source" = "https://github.com/wesfletch/toybox"


[tool.hatch.build.targets.sdist]
exclude = [
    "test/*"
]

[tool.hatch.build.targets.sdist.force-include]
"src/toybox_core/scripts/server.py" = "toybox_core/scripts/server.py"

[tool.hatch.build.targets.wheel.force-include]
"src/toybox_core/scripts/server.py" = "toybox_core/scripts/server.py"

[project.scripts]
tbx-server = "toybox_core.scripts.server:main"
tbx-echo = "toybox_core.scripts.echo:main"
tbx-publish = "toybox_core.scripts.publish:main"